Update id type

+++ backend/app/models/app_model.go
@@ -2,7 +2,7 @@ package models
 
 // User struct to describe App object.
 type App struct {
-	ID          int    `json:"id" gorm:"primarykey"`
+	ID          uint64 `json:"id" gorm:"primarykey"`
 	AppName     string `json:"app_name" gorm:"unique" validate:"required,lte=255"`
 	ApiKey      string `json:"api_key"`
 	SecretKey   string `json:"secret_key"`
